backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
motor==3.6.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.12
pydantic-settings==2.6.0
httpx==0.27.2
pytest==8.3.3

